# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "librarium API",
        "description": "API Documentation",
        "contact": {
            "email": "domenoer@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/authors": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "Return authors collection",
                "description": "Display a listing of the resource.",
                "operationId": "displayAuthors",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of records to retrieve",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return authors collection successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": [
                                                {
                                                    "id": "3",
                                                    "firstName": "Bob",
                                                    "lastName": "Fanger",
                                                    "bio": "Est aperiam optio dolor id et earum. Quasi rem nobis quibusdam doloremque illo. Exercitationem explicabo temporibus quibusdam ut."
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Author"
                ],
                "summary": "Create new author",
                "description": "Store author",
                "operationId": "storeAuthor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Mike"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Fireman"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Est aperiam optio dolor id et earum..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": "3",
                                    "firstName": "Bob",
                                    "lastName": "Fanger",
                                    "bio": "Est aperiam optio dolor id et earum. Quasi rem nobis quibusdam doloremque illo. Exercitationem explicabo temporibus quibusdam ut."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated"
                    }
                }
            }
        },
        "/api/authors/{id}": {
            "get": {
                "tags": [
                    "Author"
                ],
                "summary": "Show author",
                "description": "Display author",
                "operationId": "showAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return author successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": "3",
                                    "firstName": "Bob",
                                    "lastName": "Fanger",
                                    "bio": "Est aperiam optio dolor id et earum. Quasi rem nobis quibusdam doloremque illo. Exercitationem explicabo temporibus quibusdam ut."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Author"
                ],
                "summary": "Update author",
                "description": "Update the specified resource in storage.",
                "operationId": "updateAuthor",
                "parameters": [
                    {
                        "parameter": "update",
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 3
                        }
                    }
                ],
                "requestBody": {
                    "request": "update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Mike"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Fireman"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Est aperiam optio dolor id et earum..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": "3",
                                    "firstName": "Bob",
                                    "lastName": "Fanger",
                                    "bio": "Est aperiam optio dolor id et earum. Quasi rem nobis quibusdam doloremque illo. Exercitationem explicabo temporibus quibusdam ut."
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Author"
                ],
                "summary": "Delete author",
                "description": "Remove author.",
                "operationId": "deleteAuthor",
                "parameters": [
                    {
                        "parameter": "delete",
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "User not authenticated"
                    }
                }
            }
        },
...
    },
...
}